+++ Handlers/LeaTranslationHandler.hpp
@@ -61,7 +61,7 @@ private:
                                   .Mode(ZYDIS_MACHINE_MODE_LEGACY_32)
                                   .Mnemonic(ZYDIS_MNEMONIC_MOV)
                                   .Operand(ZYDIS_OPERAND_TYPE_MEMORY)
-                                  .Mem((uint32_t) (&additionalContext
+                                  .Mem((uint32_t)(uintptr_t)(&additionalContext
                                           .GetRegister(firstDownRegister)), 4)
                                   .FinishOperand()
                                   .Operand(ZYDIS_OPERAND_TYPE_IMMEDIATE)
